WIP: Refactor Stats computation
More view refactoring is likely necessary. A PStatsView might just want
to accept the dataclass, if we even see a need to maintain that view
protocol.
