+++ src/ndv/models/__init__.py
@@ -3,7 +3,6 @@
 from ._array_display_model import ArrayDisplayModel
 from ._data_display_model import DataDisplayModel
 from ._lut_model import LUTModel
-from ._stats_model import StatsModel
 from .data_wrappers._data_wrapper import DataWrapper
 
 __all__ = [
@@ -11,5 +10,4 @@ __all__ = [
     "DataDisplayModel",
     "DataWrapper",
     "LUTModel",
-    "StatsModel",
 ]
